Updated canary, changed event system
