+++ canary/tcp_handlers.go
@@ -65,3 +65,5 @@ func EventHTTP(sourceIP net.IP, method, uri, proto string, headers http.Header) 
 		"http.user-agent":   headers.Get("User-Agent"),
 	})
 }
+
+// port 139 -> http://s11.invisionfree.com/dongsongbang/ar/t170.htm
